+++ benchmark/src/test/scala/com/evolutiongaming/kafka/journal/replicator/BatchBenchmark.scala
@@ -20,7 +20,7 @@ import java.util.concurrent.TimeUnit
 class BatchBenchmark {
 
   @Benchmark
-  def original(blackhole: Blackhole): Unit = {
+  def original4_1_0(blackhole: Blackhole): Unit = {
     val of = Batch_4_1_0.of
     callAll[Batch_4_1_0](blackhole, of)
   }
@@ -32,15 +32,15 @@ class BatchBenchmark {
   }
 
   @Benchmark
-  def alternative4_1_2(blackhole: Blackhole): Unit = {
-    val of = Batch_4_1_2.of
-    callAll[Batch_4_1_2](blackhole, of)
+  def alternativeWithAggressiveReshuffling(blackhole: Blackhole): Unit = {
+    val of = Batch_Alternative_with_Aggressive_Reshuffling.of
+    callAll[Batch_Alternative_with_Aggressive_Reshuffling](blackhole, of)
   }
 
   @Benchmark
-  def alternative4_1_2_df(blackhole: Blackhole): Unit = {
-    val of = Batch_4_1_2_df.of
-    callAll[Batch_4_1_2_df](blackhole, of)
+  def alternative4_1_2(blackhole: Blackhole): Unit = {
+    val of = Batch_4_1_2.of
+    callAll[Batch_4_1_2](blackhole, of)
   }
 
   def callAll[T](blackhole: Blackhole, of: (Nel[ActionRecord[Action]]) => List[T]): Unit = {
