+++ benchmark/src/test/scala/com/evolutiongaming/kafka/journal/replicator/Batch_Alternative_with_Aggressive_Reshuffling.scala
@@ -28,10 +28,10 @@ private[journal] object Batch_Alternative_with_Aggressive_Reshuffling {
   }
 
   private final case class State(
-                                  private val purge: Option[Purge]     = None,
-                                  private val appends: Option[Appends] = None,
-                                  private val delete: Option[Delete]   = None,
-                                ) {
+    private val purge: Option[Purge]     = None,
+    private val appends: Option[Appends] = None,
+    private val delete: Option[Delete]   = None,
+  ) {
     // Expects records to be provided in reversed order, e.g., youngest first
     private def handle: ActionRecord[Action] => State = {
       case _ if this.purge.nonEmpty => // ignore all actions before `Purge`
